+++ CHANGELOG.md
@@ -15,6 +15,7 @@ All notable changes to this project will be documented in this file.
 - Support for installing java in FFDec windows installer
 
 ### Changed
+- Making internal flash viewer a default viewer + move FP option switch to Advanced settings / others
 - Increased scrolling speed in folder preview
 - Changed /bin/bash to /usr/bin/env bash
 - Building does not require Naashorn - uses Beanshell instead